requirements.txt
pandas
numpy
tensorflow
scikit-learn
matplotlib
seaborn
streamlit
plotly
Pillow
